refactor: separate json utils package
