+++ pkg/utils/json/utils.go
@@ -1,4 +1,4 @@
-package utils
+package json
 
 // JoinPatches joins array of serialized JSON patches to the single JSONPatch array
 func JoinPatches(patches [][]byte) []byte {
@@ -6,7 +6,6 @@ func JoinPatches(patches [][]byte) []byte {
 	if len(patches) == 0 {
 		return result
 	}
-
 	result = append(result, []byte("[\n")...)
 	for index, patch := range patches {
 		result = append(result, patch...)
